+++ modules/admin-ui/src/main/webapp/scripts/modules/events/controllers/bulkDeleteController.js
@@ -23,31 +23,14 @@
 // Controller for all event screens.
 angular.module('adminNg.controllers')
 .controller('BulkDeleteCtrl', ['$scope', 'Modal', 'FormNavigatorService', 'Table', 'Notifications',
-  'BulkDeleteResource', 'NewEventProcessing', 'TaskResource', 'decorateWithTableRowSelection',
-  'SeriesHasEventsResource', 'SeriesConfigurationResource',
-  function ($scope, Modal, FormNavigatorService, Table, Notifications, BulkDeleteResource, NewEventProcessing,
-    TaskResource, decorateWithTableRowSelection, SeriesHasEventsResource, SeriesConfigurationResource) {
+  'BulkDeleteResource', 'decorateWithTableRowSelection', 'SeriesHasEventsResource', 'SeriesConfigurationResource',
+  function ($scope, Modal, FormNavigatorService, Table, Notifications, BulkDeleteResource,
+    decorateWithTableRowSelection, SeriesHasEventsResource, SeriesConfigurationResource) {
 
-    var hasPublishedElements = function (currentEvent) {
-          var publicationCount = 0;
-          angular.forEach(currentEvent.publications, function() {
-            publicationCount++;
-          });
-
-          return publicationCount > 0 && currentEvent.publications[0].id != 'engage-live';
-        },
-        tableByPublicationStatus = function(isPublished) {
+    var createTable = function() {
           var result = {
             allSelected: true,
-            rows: (function () {
-              var result = [];
-              angular.forEach($scope.rows, function (row){
-                if (hasPublishedElements(row) === isPublished && row.selected) {
-                  result.push(row);
-                }
-              });
-              return result;
-            })()
+            rows: $scope.rows
           };
           decorateWithTableRowSelection(result);
           return result;
@@ -66,13 +49,7 @@ angular.module('adminNg.controllers')
         };
     $scope.rows = Table.copySelected();
     $scope.allSelected = true; // by default, all records are selected
-    $scope.currentForm = 'deleteForm'; // By default start on the delete form
-    $scope.processing = NewEventProcessing.get('delete-event');
-    $scope.published = tableByPublicationStatus(true);
-    $scope.unpublished = tableByPublicationStatus(false);
-    $scope.navigateTo = function (targetForm, currentForm, requiredForms) {
-      $scope.currentForm = FormNavigatorService.navigateTo(targetForm, currentForm, requiredForms);
-    };
+    $scope.table = createTable();
 
     var getSelectedSeriesIds = function () {
       var result = [];
@@ -85,22 +62,10 @@ angular.module('adminNg.controllers')
     };
 
     $scope.valid = function () {
-      var selectedCount;
       if (Table.resource.indexOf('series') >= 0) {
-        selectedCount = 0;
-        angular.forEach($scope.rows, function (row) {
-          if (row.selected) {
-            selectedCount++;
-          }
-        });
-        return selectedCount > 0;
+        return getSelectedSeriesIds().length > 0;
       } else {
-        selectedCount = countSelected($scope.unpublished.rows) + countSelected($scope.published.rows);
-        if (countSelected($scope.published.rows) > 0) {
-          return angular.isDefined($scope.processing.ud.workflow.id) && selectedCount > 0;
-        } else {
-          return selectedCount > 0;
-        }
+        return countSelected($scope.table.rows) > 0;
       }
     };
 
@@ -129,7 +94,7 @@ angular.module('adminNg.controllers')
         if (Table.resource.indexOf('series') >= 0) {
           deleteIds = getSelectedSeriesIds();
         } else {
-          angular.forEach(getSelected($scope.unpublished.rows), function (row) {
+          angular.forEach(getSelected($scope.table.rows), function (row) {
             deleteIds.push(row.id);
           });
         }
@@ -150,27 +115,6 @@ angular.module('adminNg.controllers')
             Modal.$scope.close();
           });
         }
-        if (Table.resource.indexOf('series') < 0 && countSelected($scope.published.rows) > 0) {
-          var retractEventIds = [], payload;
-          angular.forEach($scope.getPublishedEvents(), function (row) {
-            retractEventIds.push(row.id);
-          });
-          if (retractEventIds.length > 0) {
-            resetSubmitButton = false;
-            var configuration = $scope.processing.ud.workflow.selection.configuration;
-            var finalConfiguration = {};
-            // We have to duplicate the configuration for each event, since the "bulk start task"
-            // event wants per-event configuration since MH-12826.
-            for (var i = 0; i < retractEventIds.length; i++) {
-              finalConfiguration[retractEventIds[i]] = configuration;
-            }
-            payload = {
-              workflow: $scope.processing.ud.workflow.id,
-              configuration: finalConfiguration
-            };
-            TaskResource.save(payload, $scope.onSuccess, $scope.onFailure);
-          }
-        }
         if (resetSubmitButton) {
           // in this case, no callback would ever set submitButton to false again
           $scope.submitButton = false;
@@ -179,44 +123,7 @@ angular.module('adminNg.controllers')
 
     };
 
-    $scope.onSuccess = function () {
-      $scope.submitButton = false;
-      $scope.close();
-      Notifications.add('success', 'TASK_CREATED');
-    };
-
-    $scope.onFailure = function () {
-      $scope.submitButton = false;
-      $scope.close();
-      Notifications.add('error', 'TASK_NOT_CREATED', 'global', -1);
-    };
-
-    $scope.getPublishedEvents = function () {
-      return getSelected($scope.published.rows);
-    };
-
-    $scope.getUnpublishedEvents = function () {
-      return getSelected($scope.unpublished.rows);
-    };
-
-    $scope.toggleAllUnpublishedEvents = function () {
-      angular.forEach($scope.rows, function (row) {
-        if (!$scope.hasPublishedElements(row)) {
-          row.selected = $scope.events.unpublished.selected;
-        }
-      });
-    };
-
-    if (Table.resource.indexOf('series') < 0) {
-      $scope.events = {};
-      $scope.events.published = {};
-      $scope.events.published.has = $scope.published.rows.length > 0;
-      $scope.events.published.selected = true;
-      $scope.events.unpublished = {};
-      $scope.events.unpublished.has = $scope.unpublished.rows.length > 0;
-      $scope.events.unpublished.selected = true;
-    }
-    else {
+    if (Table.resource.indexOf('series') >= 0) {
       SeriesConfigurationResource.get(function (data) {
         $scope.deleteSeriesWithEventsAllowed = data.deleteSeriesWithEventsAllowed;
       });