+++ modules/admin-ui/src/main/webapp/scripts/shared/resources/eventsResource.js
@@ -93,6 +93,9 @@ angular.module('adminNg.resources')
         return row;
       });
 
-    }}
+    }},
+  delete: {method: 'DELETE', isArray: false, cancellable: false, transformResponse: function(_, __, status) {
+    return {status: status};
+  }}
   });
 }]);